Moving unit definitions to unit.py
